% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_data.R
\docType{data}
\name{calib_data}
\alias{calib_data}
\title{qPCR Calibration Curve Data}
\format{
A \code{\link{data.frame}} containing 4 columns.
\describe{
\item{Target}{unique identifier(s) for calibration curve }
\item{SQ}{copy number for each serial dilution }
\item{rep.no}{qPCR replicate identifier }
\item{Cq}{qPCR Cq(Ct) value }
}
}
\usage{
calib_data
}
\description{
Example data from qPCR serial dilution of known DNA construct.
}
\keyword{datasets}
