% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_lod.R
\name{calib_lod}
\alias{calib_lod}
\title{Calculate Model Based LOD and LOQ for qPCR eDNA Assays}
\usage{
calib_lod(data, threshold = 0.35, lod.fit = "best", loq.fit = "best")
}
\arguments{
\item{data}{A \code{\link{data.frame}} or \code{\link{tibble}} object containing calibration curve data.}

\item{threshold}{Threshold value for LOQ calculation using coefficient of variation (CV) of Cq (Ct)
values. Default value = 0.35.}

\item{lod.fit}{Character string to specify method to estimate LOD. Default: \code{"best"}. See details for other
possible values.}

\item{loq.fit}{Character string to specify method to estimate LOQ based on CV curves. Default: \code{'best'}.
Other option include \code{'decay'}, \code{'linear'} and \code{'Pn'} (see details).}
}
\value{
A \code{\link{list}} object containing the following objects:

\code{dataSum}: A summary \code{data.frame} of the original qPCR calibration data including copy number estimate (\code{Copy.Estimate}).

\code{standardsSum}: A summary \code{data.frame} of the standards containing the following variables: number of replicates (\code{Rep}), number of detections (\code{Detects})
   and detection rate (\code{Rate}) for each standard (\code{Standards}) and Target (\code{Target}) combination.
   The mean (\code{Cq.mean}), standard deviation (\code{Cq.sd}) and coefficient of variation (\code{Cq.CV}) of
   Cq values and also the coefficient of variation of copy number (\code{Copy.CV}) is given.

\code{assaySum}: A summary \code{data.frame} of the qPCR assay containing the following variables:
\describe{
\item{R.squared:}{The R-squared value of linear regression of all standards Cq-values vs log10 of the starting quantities}
\item{Slope:}{The slope of the linear regression}
\item{Intercept:}{The intercept of the linear regression}
\item{Low.95:}{The lowest standard with at least 95 percent positive detection}
\item{LOD:}{The 95 percent limit of detection as determined by dose-response modelling}
\item{LOQ:}{The limit of quantification as determined by decay modelling, using the user-selected CV threshold}
\item{rep2.LOD:}{The effective limit of detection if analysing in 2 replicates}
\item{rep3.LOD:}{The effective limit of detection if analysing in 3 replicates}
\item{rep4.LOD:}{The effective limit of detection if analysing in 4 replicates}
\item{rep5.LOD:}{The effective limit of detection if analysing in 5 replicates}
\item{rep8.LOD:}{The effective limit of detection if analysing in 8 replicates}
\item{LOD.model}{The optimal model when using \code{lod.fit = "best"} model selection, or the user specified model}
}

\code{LOQlist}: A \code{list} object containing LOQ model components for each Target.

\code{LODlist}: A \code{list} object containing LOD model components for each Target.
}
\description{
Calculate threshold and model based LOD and LOQ for qPCR eDNA assays based on one or
    more standard curves. This code has been modified and functionalised from the original
    by Merkes et al. (see references).
}
\details{
The \code{data} object contains data for at least one qPCR calibration curve
   usually presented as Cq (Ct) value and corresponding copy number from from a series of serial dilutions.
   The \code{data.frame} must contain the headers \code{Target}, \code{Cq} and \code{SQ}. The \code{Target}
   column must contain unique identifiers for each calibration curve. The \code{Cq} column contains the
   Cq (Ct) values and \code{SQ} contains the copy number data. Additional columns will be ignored. Non-detections
   in \code{data} should be represented as \code{NA}.

   The \code{lod.fit} argument is used to select the method to estimate LOD. Selecting "best" will automatically
   fit all models available from the package \code{drc} and perform model selection based on log
   likelihood values, Akaike's information criterion and residual variance using the \code{\link[drc]{mselect}}
   function. Information on possible models can be displayed with the \code{\link[drc]{getMeanFunctions}}
   function. For example, to fit a 4 parameter Weibull type II model use \code{lod.fit = "W2.4"}.

   The \code{loq.fit} argument is used to select the method to estimate LOQ. Selecting "best" will automatically
   select the model with lowest residual standard error. 'decay' uses the exponential decay model, 'linear'
   uses a linear model and 'Pn' fits an nth-order polynomial model where n is numerical. For example,
   \code{loq.fit = "P2"} will fit a 2nd-order polynomial model and \code{loq.fit = "P3"} will use a 3rd-order.
   Selecting "best" will test polynomial models up to 6th-order.
}
\examples{
\dontrun{
#
qpcr_lod <- calib_lod(data = calib_data, threshold = 0.35,
              lod.fit = "best", loq.fit = "best")
qpcr_lod

# decay modelling estimate for LOQ
qpcr_lod_decay <- calib_lod(data = calib_data, threshold = 0.35,
                    lod.fit = "best", loq.fit = "decay")
qpcr_lod_decay

# 4 parameter Weibull type II model fitted to estimate LOD

qpcr_lod_W2 <- calib_lod(data = calib_data, threshold = 0.35,
                 lod.fit = "W2.4", loq.fit = "best")
qpcr_lod_W2
 }


}
\references{
Merkes CM, Klymus KE, Allison MJ, Goldberg C, Helbing CC, Hunter ME, Jackson CA, Lance RF,
    Mangan AM, Monroe EM, Piaggio AJ, Stokdyk JP, Wilson CC, Richter C. (2019) Generic qPCR Limit of
    Detection (LOD) / Limit of Quantification (LOQ) calculator. R Script.
    Available at: https://github.com/cmerkes/qPCR_LOD_Calc. DOI: https://doi.org/10.5066/P9GT00GB.

    Klymus, Katy E., Christopher M. Merkes, Michael J. Allison, Caren S. Goldberg, Caren C. Helbing,
    Margaret E. Hunter, Craig A. Jackson, et al. ‘Reporting the Limits of Detection and Quantification
    for Environmental DNA Assays’. Environmental DNA 2, no. 3 (July 2020): 271–82.
    https://doi.org/10.1002/edn3.29.
}
