% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_plot.R
\name{calib_plot}
\alias{calib_plot}
\title{qPCR Calibration Plot}
\usage{
calib_plot(data, target, lod = NULL, robust = FALSE, ...)
}
\arguments{
\item{data}{A \code{\link{tibble}} or \code{\link{data.frame}} containing calibration curve data.}

\item{target}{A character string containing a unique identifier for the target calibration curve to be plotted.}

\item{lod}{A vector or \code{tibble} or \code{data.frame} specifying LOD and optionally LOQ values to plot.}

\item{robust}{A logical value indicating whether the fitted model should exclude standards with less than 50\%
detections. Default = \code{FALSE}.}

\item{...}{Placeholder for further arguments that might be needed by future implementations.}
}
\value{
A \code{\link{ggplot2}} object
}
\description{
Create qPCR calibation plots.
}
\details{
The \code{data} object contains data for at least one qPCR calibration curve
   usually presented as Cq (Ct) value and corresponding copy number from from a series of serial dilutions.
   The \code{data.frame} must contain the headers \code{Target}, \code{Cq} and \code{SQ}. The \code{Target}
   column must contain unique identifiers for each calibration curve. The \code{Cq} column contains the
   Cq (Ct) values and \code{SQ} contains the copy number data. Additional columns will be ignored.

   The \code{target} argument takes a single character string specifying the target calibration curve
   to plot.

   The \code{lod} argument can be supplied as a vector or as a \code{data.frame} or \code{tibble}. If
   supplied as a vector then a single LOD value or optionally an LOD and LOQ value can be included. If both LOD
   and LOQ values are specified then LOD much be supplied first in the vector (i.e. \code{c(1.5, 4.1)}).
   If supplied as a \code{data.frame} or \code{tibble} the object must contain the headers \code{Targets},
   and \code{lod} and with an optional \code{loq} header specifying the calibration curve target, the lod and loq
   values respectively.

   Non-detections in \code{data} should be represented as \code{NA}.
}
\examples{
\dontrun{
require(tibble)
# lod as tibble
lod_data <- tibble(Target = "706", lod = 1.5, loq = 4.3)
calib_plot(calib_data, target = "706", lod = lod_data)

# lod as vector
lod_data <- c(1.5, 4.3)
calib_plot(calib_data, target = "706", lod = lod_data)

# loq missing and robust = TRUE
lod_data <- tibble(Target = "706", lod = 1.5)
calib_plot(calib_data, target = "706", lod = lod_data, robust = TRUE)
}

}
