% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_plot_all.R
\name{calib_plot_all}
\alias{calib_plot_all}
\title{Plot Multiple qPCR Calibration Plots}
\usage{
calib_plot_all(calib_df, ...)
}
\arguments{
\item{calib_df}{A \code{\link{tibble}} or \code{\link{data.frame}} containing calibration curve data.}

\item{...}{Further arguments to be passed to the \code{\link{calib_plot}} function.}
}
\value{
A \code{\link{list}} of \code{\link{ggplot2}} objects.
}
\description{
Create and plot multiple qPCR calibration plots.
}
\details{
The \code{data} object contains data for at least one qPCR calibration curve
   usually presented as Cq (Ct) value and corresponding copy number from from a series of serial dilutions.
   The \code{data.frame} must contain the headers \code{Target}, \code{Cq} and \code{SQ}. The \code{Target}
   column must contain unique identifiers for each calibration curve. The \code{Cq} column contains the
   Cq (Ct) values and \code{SQ} contains the copy number data. Additional columns will be ignored.

   Non-detections in \code{data} should be represented as \code{NA}.

   LOD and LOQ values must be supplied as a \code{\link{data.frame}} or \code{\link{tibble}} object.
}
\examples{
\dontrun{
require(tibble)

# basic plots
calib_plot_all(calib_data)

# lod and loq as tibble
lod_data <- tibble(Target = c("706", "10720201", "90720202" ),
              lod = c(1.5, 4, 10), loq = c(4.3, 10, 20))
calib_plot_all(calib_data, lod = lod_data)

# Robust model fit
calib_plot_all(calib_data, lod = lod_data, robust = TRUE)
}

}
