% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_predict.R
\name{calib_predict}
\alias{calib_predict}
\title{Predict Copy Number From Ct Values Using qPCR Standard Curves}
\usage{
calib_predict(calib_df, ct_df, ...)
}
\arguments{
\item{calib_df}{A \code{\link{data.frame}} or \code{\link{tibble}} object containing calibration curve data.}

\item{ct_df}{A \code{\link{data.frame}} or \code{\link{tibble}} object containing Ct values from environmental samples.}

\item{...}{Placeholder for further arguments that might be needed by future implementations.}
}
\value{
A \code{\link{tibble}} object containing original data in \code{ct_df} with back transformed
    copy number predictions (\code{CN.back}) and associated standard errors (\code{CN.back.se}).
}
\description{
Predicts copy number (CN) from Ct (Cq) values from one
    or more standard curves using inverse prediction.
}
\details{
The \code{calib_df} object contains data for at least one qPCR calibration curve
    usually presented as Cq (Ct) value and corresponding copy number from from a series of serial dilutions.
    The \code{data.frame} must contain the headers \code{Target}, \code{Cq} and \code{SQ}. The \code{Target}
    column must contain unique identifiers for each calibration curve. The \code{Cq} column contains the
    Cq (Ct) values and \code{SQ} contains the copy number data. Additional columns will be ignored.

    The \code{ct_df} data.frame object usually contains qPCR Cq (Ct) values from sample data. The data.frame must
    contain the headers \code{calib.curve} and \code{Ct.value}. The \code{calib.curve} column must contain
    at least one unique calibration curve identifier corresponding to those in \code{calib_df}. Any identifiers
    found in \code{ct_df} but not in \code{calib_df} will be ignored and a \code{warning} displayed. The
    \code{Ct.value} column should contain the Cq (Ct) values from environmental samples. Any additional columns
    in the data.frame will remain unchanged and be included in the returned object.

    Non-detections in either the \code{calib_df} or \code{ct_df} \code{data.frames} should be represented as \code{NA}.

    Copy number is predicted from Cq (Ct) values by fitting a linear model (\code{\link{lm}}) to data from each
    calibration curve and then using the \code{\link{inverse.predict}} function from the \code{\link{chemCal-package}}
    package.
}
\note{
Copy number values in \code{calib_df} should not be log transformed prior to using this function.
}
\examples{
\dontrun{
my_pred <- calib_predict(calib_df = calib_data, ct_df = field_data)
}
}
\references{
Massart, L.M, Vandenginste, B.G.M., Buydens, L.M.C., De Jong, S., Lewi, P.J., Smeyers-Verbeke, J. (1997)
     Handbook of Chemometrics and Qualimetrics: Part A, p. 200.
}
