% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_stats.R
\name{calib_stats}
\alias{calib_stats}
\title{Fit and Summarise qPCR Calibration Curves}
\usage{
calib_stats(data, type = "effects", ...)
}
\arguments{
\item{data}{A \code{\link{tibble}} or \code{\link{data.frame}} containing data for
one or more calibration curves.}

\item{type}{The level of detail to return. \code{"effects"} returns slope and intercept estimates, \code{"r2"}
returns the \code{R^2} of the model fits.}

\item{...}{Placeholder for further arguments that might be needed by future implementations.}
}
\value{
A \code{tibble} containing summary statistics of model fit for each calibration curve.
}
\description{
Fit and summarise qPCR calibration curves.
}
\details{
The calibration curves are fitted using a linear model with the \code{\link{lm}} function. \code{SQ}
    values are \code{\link{log10}} transformed prior to model fitting.

    The \code{data} object contains data for at least one qPCR calibration curve
    usually presented as Cq (Ct) value and corresponding copy number from from a series of serial dilutions.
    The \code{data} object must contain the headers \code{Target}, \code{Cq} and \code{SQ}. The \code{Target}
    column must contain unique identifiers for each calibration curve. The \code{Cq} column contains the
    Cq (Ct) values and \code{SQ} contains the copy number data. Additional columns will be ignored.

    The \code{"type"} argument can only take the character strings \code{"effects"} or \code{"r2"}.
    Specifying \code{"effects"} will return the slope and intercept estimates, standard error for each estimate,
    the test statistic and p.value for each calibration curve. Specifying \code{"r2"} returns the \code{R^2}
    for each model.

    Non-detections in \code{data} should be represented as \code{NA}.
}
\examples{
\dontrun{
sum_stats <- calib_stats(calib_data, type = "effects")
sum_stats <- calib_stats(calib_data, type = "r2")
}
}
