% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_data.R
\docType{data}
\name{field_data}
\alias{field_data}
\title{qPCR Field Data}
\format{
A \code{\link{data.frame}} containing 6 columns
\describe{
\item{month}{the month the water sample was collected }
\item{sampling.site}{the site the water sample was collected }
\item{filter.no}{water sample identifier }
\item{rep.no}{qPCR replicate identifier }
\item{Ct.value}{qPCR Cq (Ct) value }
\item{calib.curve}{calibration curve identifier }
}
}
\usage{
field_data
}
\description{
Example data from qPCR analysis of water samples.
}
\keyword{datasets}
