% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_calib.R
\name{theme_calib}
\alias{theme_calib}
\title{ggplot2 Theme For Calibration Curve Plots}
\usage{
theme_calib(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}

\item{base_line_size}{base_line_size}

\item{base_rect_size}{base_rect_size}
}
\value{
A \code{\link{ggplot2}} object
}
\description{
ggplot2 Theme For Calibration Curve Plots
}
\examples{
\dontrun{
theme_calib
}
}
